requirements.txt
appdirs==1.4.3
click==6.7
Django==1.10.6
django-cors-headers==2.0.2
djangorestframework==3.6.2
ijson==2.3
nltk==3.2.2
packaging==16.8
psycopg2==2.7.1
pyparsing==2.2.0
python-dotenv==0.6.3
requests==2.13.0
six==1.10.0
